[FEAT] Add user profile model and form
The user's profile is created on an after_create hook on the user
model.  The profile isn't fully implemented yet. I still need to hook
up the different agent types into a radio button, and the profile
image.  However, this gives me a way to move forward with a few of
the other stories until S3 is ready, so I'm checking it in.