[CHG] Update to logic around who gets notified
If a user is only a buyer's agent, they should not receive notification
of new showings. In the future, I'll take away the ability to fill out
a map for only buyer's agents, as that also doens't make sense, but
even so, only notify those user's who want to do showings.  Also, don't
notify the user making the showing.