[CHG] Updates to login logic
The user should be taken to the users_buyers_requests_path if they
are a buyer's agent, or "both".  They go to
users_showing_appointments_path if they are only a showing agent.
Also, if the user has requested a specific page, but were blocked by
having to login, make sure to take them to that page upon successful
login.