[CHG] Better error handling around SMS numbers
We were requiring a phone number before, but that phone number could
have been almost any format.  Since we're US only, and we need to
send text messages to the phone1 value, requiring the phone to be
exactly 10 digits is where we landed, for now.  Note: the code will
strip whitespace and non-numeric values before validating, so a value
of (303) 333-4444 is valid.