[REFAC] Fix in_bounding_box scope on Showing
I had the in_bounding_box method as a class method that was not
efficient, because I couldn't figure out how to get the
within_bounding_box method from Address (provided by Geocoder) to work
on Showings instead.  I ended up copying the code and modifying it to
work on Showing instead, which cleans up all kinds of code.