[FEAT] Impl Stripe - part 1 credit card payments
This commit is the start to the Stripe related code, though not
complete.  It allows a user to enter credit card details, as well
as logic around when they are presented/forced to enter cc details
before moving on.

* A showing agent must have a credit card on file to create a showing.
* A "both" agent must also have a credit card on file before creating
  a showing.
* A buyers agent must have a bank key on file before accepting a
  showing.

Note that no credit card, or bank details are actually stored, only
the Stripe token.