[FEAT] Impl Preferred Agent functionality
An email is now allowed when creating a new showing.  If the email
matches a user in the system, that user is given a 10 minute head
start over other matching showing agents to accept the showing.  If
the email isn't in the system, they are sent an invitation and the
showing continues as ususal.

Notifications are also sent to the buyer's agent whether their
preferred agent accepted the showing or not.