[BUG] Prefrd agent email not staying, when errors
If there are errors on the new showing page, the preferred showing
agent was incorrect, either not saving (when no matching user), or
showing as a User.to_s (when matching email).

This commit switches the preferred_agent_email to be a virtual attr,
so that the data persists between requests.