[BUG] Fix after_update bug sending welcome SMS
In production the delay of database transaction commits is slow enough
that the welcome SMS and email could go out before the db has committed
which causes an error.

The fix was to change from after_update to after_commit, but that also
fires on first create, which is not when we want to send the SMS. Also,
rspec messes with the transactions, so we need to explicitly call to
run the after_commit callback.