[REFAC] Change cron to call a master process
Cron needs to fire up the whole rails environment to kick off a method
on a class.  Since we have more and more events, this was happening
multiple times per minute.  This change makes it so that cron only
has to boot up 1 rails environment, which then fires off the commands
from there, hopefully saving a bunch of memory allocation.